requirements.txt
Django==2.2.15
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
